Move Cors config
