+++ src/main/java/com/tungstun/security/config/BarApiWebSecurityConfig.java
@@ -4,13 +4,20 @@ import com.tungstun.security.config.filter.JwtAuthorizationFilter;
 import com.tungstun.security.domain.jwt.JwtValidator;
 import org.apache.commons.lang3.ArrayUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-public class BarApiWebSecurityConfig extends WebSecurityConfigurerAdapter implements WebMvcConfigurer {
+@Configuration
+@EnableWebSecurity
+@EnableGlobalMethodSecurity(
+        prePostEnabled = true,
+        securedEnabled = true,
+        jsr250Enabled = true)
+public class BarApiWebSecurityConfig extends WebSecurityConfigurerAdapter{
     private static final String LOGIN_PATH = "/api/authenticate";
     private static final String LOGIN_REFRESH_PATH = "/api/authenticate/refresh";
     private static final String REGISTER_PATH = "/api/register";
@@ -32,16 +39,6 @@ public class BarApiWebSecurityConfig extends WebSecurityConfigurerAdapter implem
 
     @Autowired
     private JwtValidator validator;
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("*")
-                .allowedHeaders("*")
-                .allowedMethods("*")
-                .exposedHeaders("*");
-        WebMvcConfigurer.super.addCorsMappings(registry);
-    }
 
     @Override
     protected void configure(HttpSecurity http) throws Exception {
